// frontend/api/models/HouseListItem.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { HouseStatus } from './HouseStatus';
export type HouseListItem = {
    id?: number;
    cover_image?: string;
    layout?: string;
    address_summary?: string;
    monthly_rent?: number;
    area?: number;
    status?: HouseStatus;
    status_label?: string;
    created_at?: string;
};